CAMERA SHAKE
Camera shake can be added with camera_shake(). Remember to only run this function once per action since running it in a loop will stack it.

Also fixed a draw problem with the hearts.

Removed the need to have the frame in each function. Just use current_draw_frame instead.